requirements.txt
aiofiles==0.6.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fastapi==0.62.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.4
google-cloud-datastore==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.34.0
h11==0.11.0
idna==2.10
Jinja2==2.11.2
libcst==0.3.15
MarkupSafe==1.1.1
mypy-extensions==0.4.3
optional.py==1.1.0
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
uvicorn==0.12.3
